Overhaul FLATDATA for char arrays only
This splits the FLATDATA macro and its associated CFlatData wrapper into:
* Support for serializing/deserializing char arrays directly, without any
  wrappers. The FLATDATA macro just becomes a pass-through dummy, to be
  removed later, when all instances have been rewritten.
* A CharArray wrapper which works only for types that offer a .data() and .size().

This includes a patch by Russell Yanofsky to make char casting type safe.

The serialization of CSubNet is changed to serialize a bool directly rather
than though FLATDATA. This makes the serialization independent of the size
of the bool type (and will use 1 byte everywhere).